iptables: Fix iptables removal logic on bootstrap
[ upstream commit 36cdd98c7278c87c4034c7a8fbdc57b74271ecbc ]

The existing legacy rule removal logic on bootstrap removed all rules which
contains the word "cilium". While this removed Cilium relevant rules, it also
incorrectly removed rules installed by the portmap/hostport plugin if the
plugin was configured with a name that contained the string cilium.

Example CNI configuration:
```
    {
      "cniVersion": "0.3.1",
        "name": "cilium-portmap",
        "plugins": [
          {
            "type": "cilium-cni"
          },
          {
            "type": "portmap",
            "capabilities": { "portMappings": true }
          }
        ]
    }
```

Example of incorrectly removed rule:
-A CNI-HOSTPORT-DNAT -m comment --comment "dnat name: \"cilium-portmap\" id: \"95dc537b9152da5f91be3fc5692bf91592bf1871b6e61755aed2056a03e98c4f\"" -j CNI-DN-258a52f03b4b7aa8abdc5

The fix is to be more restrictive in selecting rules to remove and limit it to
rules which contain the string "CILIUM_".

Fixes: #6499

Signed-off-by: Thomas Graf <thomas@cilium.io>
Signed-off-by: Ian Vernon <ian@cilium.io>